feat: UserAttendModal Container와 서버 호출 필요한 부분 컴포넌트 분리
